chore(team): split mailbox query helpers
